Split & rename basic binding views
