+++ packages/ember-views/lib/system/render_buffer.js
@@ -514,7 +514,7 @@ _RenderBuffer.prototype = {
   element: function() {
     var content = this.innerContent();
     // No content means a text node buffer, with the content
-    // in _element. HandlebarsBoundView is an example.
+    // in _element. Ember._BoundView is an example.
     if (content === null)  {
       return this._element;
     }
